Expand and re-order for clarity
I started trying to read this doc in order to manually decode an automerge
doc (mostly for curiousity). While doing so, I had to make some educated
guesses, and spent a lot of time jumping around in the document.

To make this document easier to use for that purpose, I've significantly
expanded the "Concepts" section, re-ordered it so that the document
becomes progressively more detailed, and added / clarified documentation
around the column encodings where it was missing.

There is still a *lot* of information that would be needed to write a
compatible implementation that is still missing, but this should be at
least a step in the right direction.